+++ Demo/Pods/AMScrollingNavbar/Source/ScrollingNavigationController.swift
@@ -43,32 +43,6 @@ import UIKit
 }
 
 /**
- The direction of scrolling that a followe should follow when the navbar is collapsing.
- The raw value determines the sign of content offset depending of collapse direction.
- 
- - scrollUp: scrolling up direction
- - scrollDown: scrolling down direction
- */
-@objc public enum NavigationBarFollowerCollapseDirection: Int {
-  case scrollUp = -1
-  case scrollDown = 1
-}
-
-/**
- Wraps a view that follows the navigation bar, providing the direction that the view should follow
- */
-@objcMembers
-open class NavigationBarFollower: NSObject {
-  public weak var view: UIView?
-  public var direction = NavigationBarFollowerCollapseDirection.scrollUp
-  
-  public init(view: UIView, direction: NavigationBarFollowerCollapseDirection = .scrollUp) {
-    self.view = view
-    self.direction = direction
-  }
-}
-
-/**
  A custom `UINavigationController` that enables the scrolling of the navigation bar alongside the
  scrolling of an observed content view
  */
@@ -78,14 +52,16 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   /**
    Returns the `NavigationBarState` of the navigation bar
    */
-  open var state: NavigationBarState {
-    get {
-      if navigationBar.frame.origin.y <= -navbarFullHeight {
-        return .collapsed
-      } else if navigationBar.frame.origin.y >= statusBarHeight {
-        return .expanded
-      } else {
-        return .scrolling
+  open fileprivate(set) var state: NavigationBarState = .expanded {
+    willSet {
+      if state != newValue {
+        scrollingNavbarDelegate?.scrollingNavigationController?(self, willChangeState: newValue)
+      }
+    }
+    didSet {
+      navigationBar.isUserInteractionEnabled = (state == .expanded)
+      if state != oldValue {
+        scrollingNavbarDelegate?.scrollingNavigationController?(self, didChangeState: state)
       }
     }
   }
@@ -114,46 +90,12 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   open weak var scrollingNavbarDelegate: ScrollingNavigationControllerDelegate?
 
   /**
-   An array of `NavigationBarFollower`s that will follow the navbar
+   An array of `UIView`s that will follow the navbar
    */
-  open var followers: [NavigationBarFollower] = []
-
-  /**
-   Determines if the top content inset should be updated with the navbar's delta movement. This should be enabled when dealing with table views with floating headers.
-   It can however cause issues in certain configurations. If the issues arise, set this to false
-   
-   Defaults to `true`
-   */
-  open var shouldUpdateContentInset = true
-
-  /**
-   Determines if the navigation bar should scroll while following a UITableView that is in edit mode.
-   
-   Defaults to `false`
-   */
-  open var shouldScrollWhenTableViewIsEditing = false
-  
-  /// Holds the percentage of the navigation bar that is hidde. At 0 the navigation bar is fully visible, at 1 fully hidden. CGFloat with values from 0 to 1
-  open var percentage: CGFloat {
-    get {
-      return (navigationBar.frame.origin.y - statusBarHeight) / (-navbarFullHeight - statusBarHeight)
-    }
-  }
-  
-  /// Stores some metadata of a UITabBar if one is passed in the followers array
-  internal struct TabBarMock {
-    var isTranslucent: Bool = false
-    var origin: CGPoint = .zero
-
-    init(origin: CGPoint, translucent: Bool) {
-      self.origin = origin
-      self.isTranslucent = translucent
-    }
-  }
+  open var followers: [UIView] = []
 
   open fileprivate(set) var gestureRecognizer: UIPanGestureRecognizer?
-  fileprivate var sourceTabBar: TabBarMock?
-  fileprivate var previousOrientation: UIDeviceOrientation = UIDevice.current.orientation
+  fileprivate var sourceTabBar: UITabBar?
   var delayDistance: CGFloat = 0
   var maxDelay: CGFloat = 0
   var scrollableView: UIView?
@@ -161,7 +103,7 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   var scrollSpeedFactor: CGFloat = 1
   var collapseDirectionFactor: CGFloat = 1 // Used to determine the sign of content offset depending of collapse direction
   var previousState: NavigationBarState = .expanded // Used to mark the state before the app goes in background
-  
+
   /**
    Start scrolling
 
@@ -171,41 +113,28 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
    - parameter delay: The delay expressed in points that determines the scrolling resistance. Defaults to `0`
    - parameter scrollSpeedFactor : This factor determines the speed of the scrolling content toward the navigation bar animation
    - parameter collapseDirection : The direction of scrolling that the navigation bar should be collapsed
-   - parameter followers: An array of `NavigationBarFollower`s that will follow the navbar. The wrapper holds the direction that the view will follow
+   - parameter followers: An array of `UIView`s that will follow the navbar
    */
-  open func followScrollView(_ scrollableView: UIView, delay: Double = 0, scrollSpeedFactor: Double = 1, collapseDirection: NavigationBarCollapseDirection = .scrollDown, followers: [NavigationBarFollower] = []) {
-    guard self.scrollableView == nil else {
-      // Restore previous state. UIKit restores the navbar to its full height on view changes (e.g. during a modal preentation), so we need to restore the status once UIKit is done
-      switch previousState {
-      case .collapsed:
-        hideNavbar(animated: false)
-      case .expanded:
-        showNavbar(animated: false)
-      default: break
-      }
-      return
-    }
+  open func followScrollView(_ scrollableView: UIView, delay: Double = 0, scrollSpeedFactor: Double = 1, collapseDirection: NavigationBarCollapseDirection = .scrollDown, followers: [UIView] = []) {
     self.scrollableView = scrollableView
 
     gestureRecognizer = UIPanGestureRecognizer(target: self, action: #selector(ScrollingNavigationController.handlePan(_:)))
     gestureRecognizer?.maximumNumberOfTouches = 1
     gestureRecognizer?.delegate = self
-    gestureRecognizer?.cancelsTouchesInView = false
     scrollableView.addGestureRecognizer(gestureRecognizer!)
 
-    previousOrientation = UIDevice.current.orientation
-    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.willResignActive(_:)), name: UIApplication.willResignActiveNotification, object: nil)
-    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.didBecomeActive(_:)), name: UIApplication.didBecomeActiveNotification, object: nil)
-    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.didRotate(_:)), name: UIDevice.orientationDidChangeNotification, object: nil)
-    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.windowDidBecomeVisible(_:)), name: UIWindow.didBecomeVisibleNotification, object: nil)
+    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.willResignActive(_:)), name: NSNotification.Name.UIApplicationWillResignActive, object: nil)
+    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.didBecomeActive(_:)), name: NSNotification.Name.UIApplicationDidBecomeActive, object: nil)
+    NotificationCenter.default.addObserver(self, selector: #selector(ScrollingNavigationController.didRotate(_:)), name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
 
     maxDelay = CGFloat(delay)
     delayDistance = CGFloat(delay)
     scrollingEnabled = true
 
     // Save TabBar state (the state is changed during the transition and restored on compeltion)
-    if let tab = followers.map({ $0.view }).first(where: { $0 is UITabBar }) as? UITabBar {
-      self.sourceTabBar = TabBarMock(origin: CGPoint(x: tab.frame.origin.x, y: CGFloat(round(tab.frame.origin.y))), translucent: tab.isTranslucent)
+    if let tab = followers.first(where: { $0 is UITabBar }) as? UITabBar {
+      self.sourceTabBar = UITabBar(frame: tab.frame)
+      self.sourceTabBar?.isTranslucent = tab.isTranslucent
     }
     self.followers = followers
     self.scrollSpeedFactor = CGFloat(scrollSpeedFactor)
@@ -221,28 +150,20 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   open func hideNavbar(animated: Bool = true, duration: TimeInterval = 0.1) {
     guard let _ = self.scrollableView, let visibleViewController = self.visibleViewController else { return }
 
-    guard state == .expanded else {
-      updateNavbarAlpha()
-      return
-    }
-
-    gestureRecognizer?.isEnabled = false
-    let animations = {
-      self.scrollWithDelta(self.fullNavbarHeight, ignoreDelay: true)
-      visibleViewController.view.setNeedsLayout()
-      if self.navigationBar.isTranslucent {
-        let currentOffset = self.contentOffset
-        self.scrollView()?.contentOffset = CGPoint(x: currentOffset.x, y: currentOffset.y + self.navbarHeight)
-      }
-    }
-
-    if animated {
-      UIView.animate(withDuration: duration, animations: animations) { _ in
-        self.gestureRecognizer?.isEnabled = true
+    if state == .expanded {
+      self.state = .scrolling
+      UIView.animate(withDuration: animated ? duration : 0, animations: { () -> Void in
+        self.scrollWithDelta(self.fullNavbarHeight)
+        visibleViewController.view.setNeedsLayout()
+        if self.navigationBar.isTranslucent {
+          let currentOffset = self.contentOffset
+          self.scrollView()?.contentOffset = CGPoint(x: currentOffset.x, y: currentOffset.y + self.navbarHeight)
+        }
+      }) { _ in
+        self.state = .collapsed
       }
     } else {
-      animations()
-      gestureRecognizer?.isEnabled = true
+      updateNavbarAlpha()
     }
   }
 
@@ -255,28 +176,30 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   open func showNavbar(animated: Bool = true, duration: TimeInterval = 0.1) {
     guard let _ = self.scrollableView, let visibleViewController = self.visibleViewController else { return }
 
-    guard state == .collapsed else {
-      updateNavbarAlpha()
-      return
-    }
-
-    gestureRecognizer?.isEnabled = false
-    let animations = {
-      self.lastContentOffset = 0
-      self.scrollWithDelta(-self.fullNavbarHeight, ignoreDelay: true)
-      visibleViewController.view.setNeedsLayout()
-      if self.navigationBar.isTranslucent {
-        let currentOffset = self.contentOffset
-        self.scrollView()?.contentOffset = CGPoint(x: currentOffset.x, y: currentOffset.y - self.navbarHeight)
+    if state == .collapsed {
+      gestureRecognizer?.isEnabled = false
+      let animations = {
+        self.lastContentOffset = 0;
+        self.scrollWithDelta(-self.fullNavbarHeight, ignoreDelay: true)
+        visibleViewController.view.setNeedsLayout()
+        if self.navigationBar.isTranslucent {
+          let currentOffset = self.contentOffset
+          self.scrollView()?.contentOffset = CGPoint(x: currentOffset.x, y: currentOffset.y - self.navbarHeight)
+        }
       }
-    }
-    if animated {
-      UIView.animate(withDuration: duration, animations: animations) { _ in
+      if animated {
+        self.state = .scrolling
+        UIView.animate(withDuration: duration, animations: animations) { _ in
+          self.state = .expanded
+          self.gestureRecognizer?.isEnabled = true
+        }
+      } else {
+        animations()
+        self.state = .expanded
         self.gestureRecognizer?.isEnabled = true
       }
     } else {
-      animations()
-      gestureRecognizer?.isEnabled = true
+      updateNavbarAlpha()
     }
   }
 
@@ -298,54 +221,35 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
     scrollingEnabled = false
 
     let center = NotificationCenter.default
-    center.removeObserver(self, name: UIApplication.didBecomeActiveNotification, object: nil)
-    center.removeObserver(self, name: UIDevice.orientationDidChangeNotification, object: nil)
+    center.removeObserver(self, name: NSNotification.Name.UIApplicationDidBecomeActive, object: nil)
+    center.removeObserver(self, name: NSNotification.Name.UIDeviceOrientationDidChange, object: nil)
   }
 
   // MARK: - Gesture recognizer
 
   func handlePan(_ gesture: UIPanGestureRecognizer) {
-    if let tableView = scrollableView as? UITableView, !shouldScrollWhenTableViewIsEditing && tableView.isEditing {
-      return
-    }
-    if let superview = scrollableView?.superview {
-      let translation = gesture.translation(in: superview)
-      let delta = (lastContentOffset - translation.y) / scrollSpeedFactor
-      
-      if !checkSearchController(delta) {
+    if gesture.state != .failed {
+      if let superview = scrollableView?.superview {
+        let translation = gesture.translation(in: superview)
+        let delta = collapseDirectionFactor * (lastContentOffset - translation.y) / scrollSpeedFactor
         lastContentOffset = translation.y
-        return
-      }
-      
-      if gesture.state != .failed {
-        lastContentOffset = translation.y
+
         if shouldScrollWithDelta(delta) {
           scrollWithDelta(delta)
         }
       }
     }
-    
+
     if gesture.state == .ended || gesture.state == .cancelled || gesture.state == .failed {
       checkForPartialScroll()
       lastContentOffset = 0
     }
   }
-  
-  // MARK: - Fullscreen handling
-
-  func windowDidBecomeVisible(_ notification: Notification) {
-    showNavbar()
-  }
 
   // MARK: - Rotation handler
 
   func didRotate(_ notification: Notification) {
-    let newOrientation = UIDevice.current.orientation
-    // Show the navbar if the orantation is the same (the app just got back from background) or if there is a switch between portrait and landscape (and vice versa)
-    if (previousOrientation == newOrientation) || (previousOrientation.isPortrait && newOrientation.isLandscape) || (previousOrientation.isLandscape && newOrientation.isPortrait) {
-      showNavbar()
-    }
-    previousOrientation = newOrientation
+    showNavbar()
   }
 
   /**
@@ -382,10 +286,6 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
 
   private func shouldScrollWithDelta(_ delta: CGFloat) -> Bool {
     let scrollDelta = delta
-    // Do not hide too early
-    if contentOffset.y < ((navigationBar.isTranslucent ? -fullNavbarHeight : 0) + scrollDelta) {
-      return false
-    }
     // Check for rubberbanding
     if scrollDelta < 0 {
       if let scrollableView = scrollableView , contentOffset.y + scrollableView.frame.size.height > contentSize.height && scrollableView.frame.size.height < contentSize.height {
@@ -419,13 +319,16 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
       }
 
       // Compute the bar position
-      if frame.origin.y - scrollDelta < -navbarFullHeight {
-        scrollDelta = frame.origin.y + navbarFullHeight
+      if frame.origin.y - scrollDelta < -deltaLimit {
+        scrollDelta = frame.origin.y + deltaLimit
       }
 
       // Detect when the bar is completely collapsed
-      if frame.origin.y <= -navbarFullHeight {
+      if frame.origin.y <= -deltaLimit {
+        state = .collapsed
         delayDistance = maxDelay
+      } else {
+        state = .scrolling
       }
     }
 
@@ -447,55 +350,39 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
 
       // Detect when the bar is completely expanded
       if frame.origin.y >= statusBarHeight {
+        state = .expanded
         delayDistance = maxDelay
+      } else {
+        state = .scrolling
       }
     }
-    
+
     updateSizing(scrollDelta)
     updateNavbarAlpha()
     restoreContentOffset(scrollDelta)
-    updateFollowers()
+    updateFollowers(scrollDelta)
     updateContentInset(scrollDelta)
-    
-    let newState = state
-    if newState != previousState {
-      scrollingNavbarDelegate?.scrollingNavigationController?(self, willChangeState: newState)
-      navigationBar.isUserInteractionEnabled = (newState == .expanded)
-    }
-    previousState = newState
   }
 
   /// Adjust the top inset (useful when a table view has floating headers, see issue #219
   private func updateContentInset(_ delta: CGFloat) {
-    if self.shouldUpdateContentInset, let contentInset = scrollView()?.contentInset, let scrollInset = scrollView()?.scrollIndicatorInsets {
+    if let contentInset = scrollView()?.contentInset, let scrollInset = scrollView()?.scrollIndicatorInsets {
       scrollView()?.contentInset = UIEdgeInsets(top: contentInset.top - delta, left: contentInset.left, bottom: contentInset.bottom, right: contentInset.right)
       scrollView()?.scrollIndicatorInsets = UIEdgeInsets(top: scrollInset.top - delta, left: scrollInset.left, bottom: scrollInset.bottom, right: scrollInset.right)
     }
   }
 
-  private func updateFollowers() {
+  private func updateFollowers(_ delta: CGFloat) {
     followers.forEach {
-      guard let tabBar = $0.view as? UITabBar else {
-        let height = $0.view?.frame.height ?? 0
-        var safeArea: CGFloat = 0
-        if #available(iOS 11.0, *) {
-          // Account for the safe area for footers and toolbars at the bottom of the screen
-          safeArea = ($0.direction == .scrollDown) ? (topViewController?.view.safeAreaInsets.bottom ?? 0) : 0
-        }
-        switch $0.direction {
-        case .scrollDown:
-          $0.view?.transform = CGAffineTransform(translationX: 0, y: percentage * (height + safeArea))
-        case .scrollUp:
-          $0.view?.transform = CGAffineTransform(translationX: 0, y: -(statusBarHeight - navigationBar.frame.origin.y))
-        }
-        
+      guard let tabBar = $0 as? UITabBar else {
+        $0.transform = $0.transform.translatedBy(x: 0, y: -delta)
         return
       }
       tabBar.isTranslucent = true
-      tabBar.transform = CGAffineTransform(translationX: 0, y: percentage * tabBar.frame.height)
+      tabBar.frame.origin.y += delta * 1.5
 
       // Set the bar to its original state if it's in its original position
-      if let originalTabBar = sourceTabBar, originalTabBar.origin.y == round(tabBar.frame.origin.y) {
+      if let originalTabBar = sourceTabBar, originalTabBar.frame.origin.y == tabBar.frame.origin.y {
         tabBar.isTranslucent = originalTabBar.isTranslucent
       }
     }
@@ -542,26 +429,23 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
       delta = frame.origin.y - statusBarHeight
       let distance = delta / (frame.size.height / 2)
       duration = TimeInterval(abs(distance * 0.2))
-      scrollingNavbarDelegate?.scrollingNavigationController?(self, willChangeState: state)
+      state = .expanded
     } else {
       // Scroll up
-      delta = frame.origin.y + navbarFullHeight
+      delta = frame.origin.y + deltaLimit
       let distance = delta / (frame.size.height / 2)
       duration = TimeInterval(abs(distance * 0.2))
-      scrollingNavbarDelegate?.scrollingNavigationController?(self, willChangeState: state)
+      state = .collapsed
     }
 
     delayDistance = maxDelay
     
-    UIView.animate(withDuration: duration, delay: 0, options: UIView.AnimationOptions.beginFromCurrentState, animations: {
+    UIView.animate(withDuration: duration, delay: 0, options: UIViewAnimationOptions.beginFromCurrentState, animations: {
       self.updateSizing(delta)
-      self.updateFollowers()
+      self.updateFollowers(delta)
       self.updateNavbarAlpha()
       self.updateContentInset(delta)
-    }, completion: { _ in
-      self.navigationBar.isUserInteractionEnabled = (self.state == .expanded)
-      self.scrollingNavbarDelegate?.scrollingNavigationController?(self, didChangeState: self.state)
-    })
+    }, completion: nil)
   }
 
   private func updateNavbarAlpha() {
@@ -570,24 +454,15 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
     let frame = navigationBar.frame
 
     // Change the alpha channel of every item on the navbr
-    let alpha = 1 - percentage
+    let alpha = (frame.origin.y + deltaLimit) / frame.size.height
 
     // Hide all the possible titles
     navigationItem.titleView?.alpha = alpha
     navigationBar.tintColor = navigationBar.tintColor.withAlphaComponent(alpha)
-    navigationItem.leftBarButtonItem?.tintColor = navigationItem.leftBarButtonItem?.tintColor?.withAlphaComponent(alpha)
-    navigationItem.rightBarButtonItem?.tintColor = navigationItem.rightBarButtonItem?.tintColor?.withAlphaComponent(alpha)
-    navigationItem.leftBarButtonItems?.forEach { $0.tintColor = $0.tintColor?.withAlphaComponent(alpha) }
-    navigationItem.rightBarButtonItems?.forEach { $0.tintColor = $0.tintColor?.withAlphaComponent(alpha) }
-    if let titleColor = navigationBar.titleTextAttributes?[NSAttributedString.Key.foregroundColor] as? UIColor {
-      navigationBar.titleTextAttributes?[NSAttributedString.Key.foregroundColor] = titleColor.withAlphaComponent(alpha)
+    if let titleColor = navigationBar.titleTextAttributes?[NSAttributedStringKey.foregroundColor] as? UIColor {
+      navigationBar.titleTextAttributes?[NSAttributedStringKey.foregroundColor] = titleColor.withAlphaComponent(alpha)
     } else {
-      let blackAlpha = UIColor.black.withAlphaComponent(alpha)
-      if navigationBar.titleTextAttributes == nil {
-        navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor: blackAlpha]
-      } else {
-        navigationBar.titleTextAttributes?[NSAttributedString.Key.foregroundColor] = blackAlpha
-      }
+      navigationBar.titleTextAttributes?[NSAttributedStringKey.foregroundColor] = UIColor.black.withAlphaComponent(alpha)
     }
 
     // Hide all possible button items and navigation items
@@ -603,15 +478,7 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
     }
 
     func setAlphaOfSubviews(view: UIView, alpha: CGFloat) {
-      if let label = view as? UILabel {
-        label.textColor = label.textColor.withAlphaComponent(alpha)
-      } else if let label = view as? UITextField {
-        label.textColor = label.textColor?.withAlphaComponent(alpha)
-      } else if view.classForCoder == NSClassFromString("_UINavigationBarContentView") {
-        // do nothing
-      } else {
-        view.alpha = alpha
-      }
+      view.alpha = alpha
       view.subviews.forEach { setAlphaOfSubviews(view: $0, alpha: alpha) }
     }
 
@@ -627,17 +494,6 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
     navigationItem.rightBarButtonItem?.customView?.alpha = alpha
     navigationItem.rightBarButtonItems?.forEach { $0.customView?.alpha = alpha }
   }
-  
-  private func checkSearchController(_ delta: CGFloat) -> Bool {
-    if #available(iOS 11.0, *) {
-      if let searchController = topViewController?.navigationItem.searchController, delta > 0 {
-        if searchController.searchBar.frame.height != 0 {
-          return false
-        }
-      }
-    }
-    return true
-  }
 
   // MARK: - UIGestureRecognizerDelegate
 
@@ -647,7 +503,7 @@ open class ScrollingNavigationController: UINavigationController, UIGestureRecog
   open func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
     guard let gestureRecognizer = gestureRecognizer as? UIPanGestureRecognizer else { return true }
     let velocity = gestureRecognizer.velocity(in: gestureRecognizer.view)
-    return abs(velocity.y) > abs(velocity.x)
+    return fabs(velocity.y) > fabs(velocity.x)
   }
 
   /**